准备部署

+++ src/main/java/com/lyj/blog/service/BlogService.java
@@ -38,15 +38,13 @@ public class BlogService {
 
         //是否清除和添加elasticsearch的数据
         if (true) {
-            try {
+            boolean existIndex = ElasticseachClientUtil.existIndex("header");
+            if(existIndex){
                 ElasticseachClientUtil.deleteIndex("header");
-            } catch (IOException e) {
-                System.out.println("还未创建header索引");
             }
 
             //批量添加header数据到elasticsearch
             elasticsearchService.addHeaderBulk();
-
         }
     }
